+++ mentat/command/__init__.py
@@ -1,0 +1,3 @@
+# Make sure commands are initialized
+
+from . import commands  # noqa: F401 # type: ignore
